requirements.txt
# Core framework
fastapi==0.115.0
uvicorn==0.30.1

# LLM + LangGraph + LangChain
langchain==0.2.14
langgraph==0.1.6
langchain-core==0.2.22
langchain-google-genai==1.0.8

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2

# Testing
pytest==8.3.3
httpx==0.27.2

# Optional (for docs, API schemas)
openapi-schema-pydantic==1.2.4

# If you’re using Jupyter or notebooks for demo
ipykernel==6.29.5
